move MongoDatabaseExtensions to test project
