+++ test/JsonApiDotNetCoreMongoDbExampleTests/IntegrationTests/SparseFieldSets/ResultCapturingRepository.cs
@@ -2,7 +2,6 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using JsonApiDotNetCore.Configuration;
-using JsonApiDotNetCore.MongoDb;
 using JsonApiDotNetCore.MongoDb.Repositories;
 using JsonApiDotNetCore.Queries;
 using JsonApiDotNetCore.Resources;
